requirements.txt
google-genai
pydantic
pandas
openpyxl 
